Improve browser warnings
To simplify feature detection, we display a warning to any browsers that
(a) don't have javascript enabled, or (b) don't support ES6 modules.
Older browsers will blindly run the nomodule script and remove the `js`
class from the document, which will cause the warning to be displayed.

ES6 module support covers all the APIs we're using, except on Edge 16-18
which supports ES6 modules but not fetching `data:` URLS. See
 https://caniuse.com/cryptography,es6,es6-module,mdn-api_fetch_blob_data_support